Fix function generateField
This commit fix the Minesweeper function `generateField()`, which was
wrongly generating the Map object that describes a field. It generated
fields with height and width inverted.

There was only one test for this function, which tested the generation
of a square field. Tests for generating rectangular fiels, in both
orientations (horizontal/vertical) were added.