Standardize field size arguments
This commit standardize field size arguments across functions. Before,
there were two functions, `generateField()` and `adjacentVertices()`
which received the field size, but each one in a different format.

Now, the standard is functions that must receive the field size should
get a field size array (List([widht, height]). The dimensions were
inverted to be more semantic with the vertice coordinates: vertices have
the format [x, y], so width corresponds to the x-axis and height to the
y-axis.